+++ app/templates/Uzivatel/show.latte
@@ -37,12 +37,12 @@
     <table class="table table-nonfluid">
 
         {if $master}
-        <tr class="error"><th>Sloučen pod uživatele </th><td><a n:href="uzivatel:show, id=>$master->id">{$master->nick}</a></td></tr>
+        <tr class="error"><th>Sloučen pod uživatele </th><td><a n:href="Uzivatel:show, id=>$master->id">{$master->nick}</a></td></tr>
         <tr><td colspan="2">&nbsp;</td></tr>
         {/if}
 
         {if $slaves}
-        <tr n:foreach="$slaves as $slave" class="error"><th>Sloučený uživatel </th><td><a n:href="uzivatel:show, id=>$slave">{$slave}</a></td></tr>
+        <tr n:foreach="$slaves as $slave" class="error"><th>Sloučený uživatel </th><td><a n:href="Uzivatel:show, id=>$slave">{$slave}</a></td></tr>
         <tr><td colspan="2">&nbsp;</td></tr>
         {/if}
